# contracts/openapi/devpath-openapi.yaml
openapi: 3.0.3
info:
  title: DevPath Implemented API
  version: 0.11.0
  description: >
    Implemented contract for authenticated identity, repository synchronization, deterministic analysis,
    rule evaluation, Skill Matrix, career catalog, career readiness, deterministic recommendation,
    learning roadmap, and company catalog slices. Provider access tokens and unimplemented APIs are excluded.
servers:
  - url: http://localhost:8080
    description: Local backend
paths:
...
  /login/oauth2/code/github:
    get:
      summary: Complete GitHub OAuth login
      description: >
        Spring Security owns this callback. A successful callback provisions or resolves the internal
        user, rotates the session identifier, stores the security context in the server-side JDBC
        session, and redirects to the configured frontend origin.
      operationId: completeGitHubLogin
      tags: [Authentication]
      security: []
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
        - name: error
          in: query
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Redirect to the configured frontend success or failure location.
...
  /api/v1/integrations/github:
    delete:
      x-api-id: API-INT-003
      summary: Disconnect GitHub repository access
      operationId: disconnectGitHubIntegration
      tags: [Integration]
      security: [{sessionCookie: []}]
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
      responses:
        "200":
          description: GitHub access was removed locally and remote token revocation was attempted.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ConnectedAccountEnvelope"}
        "401":
          description: No valid session.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "403":
          description: Missing or invalid CSRF token.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "404":
          description: No active GitHub connection exists for the current user.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "503":
          description: Local credential removal could not be completed safely.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
  /api/v1/integrations/github/authorize:
    post:
      x-api-id: API-INT-001
      summary: Start GitHub App user authorization
      operationId: initiateGitHubIntegration
      tags: [Integration]
      security: [{sessionCookie: []}]
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/InitiateOAuthRequest"}
      responses:
        "200":
          description: Short-lived, session-bound authorization URL.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/OAuthAuthorizationEnvelope"}
        "401":
          description: No valid session.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "403":
          description: Missing or invalid CSRF token.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "503":
          description: GitHub integration configuration is unavailable.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
  /api/v1/integrations/github/callback:
    get:
      x-api-id: API-INT-002
      summary: Complete GitHub App user authorization
      operationId: completeGitHubIntegration
      tags: [Integration]
      security: [{sessionCookie: []}]
      parameters:
        - {name: state, in: query, required: true, schema: {type: string}}
        - {name: code, in: query, required: false, schema: {type: string}}
        - {name: error, in: query, required: false, schema: {type: string}}
      responses:
        "302":
          description: Redirect to the configured frontend with a safe connection result.
        "401":
          description: The account-linking session is unavailable.
  /api/v1/integrations/github/repositories:
    get:
      x-api-id: API-INT-004
      summary: List repositories accessible through installed GitHub Apps
      operationId: listGitHubRepositories
      tags: [Integration]
      security: [{sessionCookie: []}]
      responses:
        "200":
          description: Provider-normalized repositories accessible to the current user and app installation.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/GitHubRepositoryListEnvelope"}
        "401":
          description: No valid session.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "503":
          description: GitHub is unavailable or reconnection is required.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
  /api/v1/repositories/imports:
    post:
      x-api-id: API-INT-005
      summary: Import an accessible GitHub repository into DevPath
      operationId: importGitHubRepository
      tags: [Repository]
      security: [{sessionCookie: []}]
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ImportRepositoryRequest"}
      responses:
        "201":
          description: Canonical imported repository metadata. Duplicate imports return the existing owner-scoped resource.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/RepositoryEnvelope"}
        "400":
          description: Invalid or inaccessible provider repository reference.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "401":
          description: No valid session.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "403":
          description: Missing or invalid CSRF token.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "409":
          description: A concurrent repository registration changed the resource state.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "503":
          description: GitHub repository verification is unavailable.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
  /api/v1/repositories:
    get:
      x-api-id: API-REP-001
      summary: List imported repositories owned by the current user
      operationId: listRepositories
      tags: [Repository]
      security: [{sessionCookie: []}]
      parameters:
        - {name: limit, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
        - {name: cursor, in: query, required: false, schema: {type: string}}
        - {name: includeArchived, in: query, required: false, schema: {type: boolean, default: false}}
      responses:
        "200":
          description: Bounded owner-scoped repository page.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/RepositoryListEnvelope"}
        "400":
          description: Invalid limit or cursor.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "401":
          description: No valid session.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
  /api/v1/repositories/{repositoryId}:
    get:
      x-api-id: API-REP-002
      summary: Retrieve an imported repository owned by the current user
      operationId: getRepository
      tags: [Repository]
      security: [{sessionCookie: []}]
      parameters:
        - {name: repositoryId, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        "200":
          description: Owner-scoped repository metadata and current synchronization state.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/RepositoryEnvelope"}
        "401":
          description: No valid session.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "404":
          description: Repository not found for the current owner.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
  /api/v1/repositories/{repositoryId}/archive:
    post:
      x-api-id: API-REP-004
      summary: Archive an imported repository locally
      operationId: archiveRepository
      tags: [Repository]
      security: [{sessionCookie: []}]
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
        - {name: repositoryId, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        "200":
          description: Canonical repository metadata in the archived lifecycle state.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/RepositoryEnvelope"}
        "401":
          description: No valid session.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "404":
          description: Repository not found for the current owner.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "409":
          description: A concurrent repository lifecycle change caused a conflict.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
  /api/v1/repositories/{repositoryId}/restore:
    post:
      x-api-id: API-REP-005
      summary: Restore a locally archived repository
      operationId: restoreRepository
      tags: [Repository]
      security: [{sessionCookie: []}]
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
        - {name: repositoryId, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        "200":
          description: Canonical repository metadata in its restored lifecycle state.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/RepositoryEnvelope"}
        "400":
          description: Provider-archived repository cannot be restored locally.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "401":
          description: No valid session.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "404":
          description: Repository not found for the current owner.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
        "409":
          description: A concurrent repository lifecycle change caused a conflict.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: DEVPATH_SESSION
      description: Opaque HttpOnly server-session identifier. It is never read by browser JavaScript.
  parameters:
    CsrfHeader:
      name: X-CSRF-TOKEN
      in: header
      required: true
      description: Server-issued CSRF token. This is not an authentication credential.
      schema:
        type: string
...
  schemas:
    ApiMetadata:
      type: object
      additionalProperties: false
      required: [requestId, apiVersion, timestamp]
      properties:
        requestId:
          type: string
        apiVersion:
          type: string
          example: v1
        timestamp:
          type: string
          format: date-time
...
    ConnectedAccount:
      type: object
      additionalProperties: false
      required: [connectionId, provider, status, scopes, connectedAt, expiresAt]
      properties:
        connectionId: {type: string, format: uuid}
        provider: {type: string, enum: [GITHUB]}
        status: {type: string, enum: [ACTIVE, REVOKED]}
        scopes:
          type: array
          items: {type: string}
        connectedAt: {type: string, format: date-time}
        expiresAt: {type: string, format: date-time, nullable: true}
...
    ConnectedAccountEnvelope:
      type: object
      additionalProperties: false
      required: [data, metadata]
      properties:
        data: {$ref: "#/components/schemas/ConnectedAccount"}
        metadata: {$ref: "#/components/schemas/ApiMetadata"}
...
    InitiateOAuthRequest:
      type: object
      additionalProperties: false
    OAuthAuthorization:
      type: object
      additionalProperties: false
      required: [authorizationUrl]
      properties:
        authorizationUrl: {type: string, format: uri}
    OAuthAuthorizationEnvelope:
      type: object
      additionalProperties: false
      required: [data, metadata]
      properties:
        data: {$ref: "#/components/schemas/OAuthAuthorization"}
        metadata: {$ref: "#/components/schemas/ApiMetadata"}
    GitHubRepository:
      type: object
      additionalProperties: false
      required: [providerRepositoryId, name, fullName, owner, privateRepository, archived, defaultBranch, htmlUrl]
      properties:
        providerRepositoryId: {type: string}
        name: {type: string}
        fullName: {type: string}
        owner: {type: string}
        privateRepository: {type: boolean}
        archived: {type: boolean}
        defaultBranch: {type: string}
        htmlUrl: {type: string, format: uri}
    GitHubRepositoryList:
      type: object
      additionalProperties: false
      required: [repositories]
      properties:
        repositories:
          type: array
          items: {$ref: "#/components/schemas/GitHubRepository"}
    GitHubRepositoryListEnvelope:
      type: object
      additionalProperties: false
      required: [data, metadata]
      properties:
        data: {$ref: "#/components/schemas/GitHubRepositoryList"}
        metadata: {$ref: "#/components/schemas/ApiMetadata"}
    ImportRepositoryRequest:
      type: object
      additionalProperties: false
      required: [providerRepositoryId]
      properties:
        providerRepositoryId: {type: string, pattern: "^[1-9][0-9]{0,63}$"}
    Repository:
      type: object
      additionalProperties: false
      required: [repositoryId, providerRepositoryId, name, fullName, owner, visibility, defaultBranch, providerArchived, lifecycle, syncStatus, htmlUrl, discoveredAt, lastSyncedAt, currentSnapshotId]
      properties:
        repositoryId: {type: string, format: uuid}
        providerRepositoryId: {type: string}
        name: {type: string, maxLength: 100}
        fullName: {type: string, maxLength: 255}
        owner: {type: string, maxLength: 255}
        visibility: {type: string, enum: [PUBLIC, PRIVATE]}
        defaultBranch: {type: string, maxLength: 255}
        providerArchived: {type: boolean}
        lifecycle: {type: string, enum: [DISCOVERED, ACTIVE, ARCHIVED, DELETED_EXTERNALLY]}
        syncStatus: {type: string, enum: [NOT_SYNCED, SYNCHRONIZED, FAILED]}
        htmlUrl: {type: string, format: uri}
        discoveredAt: {type: string, format: date-time}
        lastSyncedAt: {type: string, format: date-time, nullable: true}
        currentSnapshotId: {type: string, format: uuid, nullable: true}
    RepositoryEnvelope:
      type: object
      additionalProperties: false
      required: [data, metadata]
      properties:
        data: {$ref: "#/components/schemas/Repository"}
        metadata: {$ref: "#/components/schemas/ApiMetadata"}
    RepositoryList:
      type: object
      additionalProperties: false
      required: [repositories, limit, nextCursor, totalCount]
      properties:
        repositories:
          type: array
          items: {$ref: "#/components/schemas/Repository"}
        limit: {type: integer, minimum: 1, maximum: 100}
        nextCursor: {type: string, nullable: true}
        totalCount: {type: integer, format: int64, minimum: 0}
    RepositoryListEnvelope:
      type: object
      additionalProperties: false
      required: [data, metadata]
      properties:
        data: {$ref: "#/components/schemas/RepositoryList"}
        metadata: {$ref: "#/components/schemas/ApiMetadata"}
...
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error, metadata]
      properties:
        error:
          $ref: "#/components/schemas/ApiError"
        metadata:
          $ref: "#/components/schemas/ApiMetadata"